Extract transcript message controls
