+++ docs/DECISIONS.md
@@ -129,5 +129,6 @@
 - MCP tool schema metadata is summarized, not replayed as arbitrary JSON. The probe stores bounded tool descriptors with name, description, required/optional argument names, and compact type summaries from `inputSchema`. Extensions UI and the dynamic `host.mcp.call` description consume the same descriptors so users and models see required arguments without generating one QuillCode tool per MCP tool or persisting unbounded schemas.
 - MCP resource and prompt discovery now feeds read-only agent tools. If a server advertises `resources` or `prompts` in `initialize`, QuillCode performs bounded optional `resources/list` and `prompts/list` probes, stores resource display names plus URIs, surfaces the counts and first names in Extensions, and ignores probe failures so an otherwise healthy MCP server is not hidden. Ready servers expose `host.mcp.resource.read` and `host.mcp.prompt.get` only for advertised resource identifiers and prompt names, keeping resource/prompt consumption read-risk and separate from append-risk generic `host.mcp.call`.
 - Shared visual primitives live in `QuillCodeDesignSystem.swift`, not in individual feature views or the workspace shell. Palette constants, hit-target metrics, `0.96` press feedback, animation helpers, reusable surfaces, and image outlines should be extended there first so extracted SwiftUI controls stay visually consistent without depending on `WorkspaceSwiftUIView.swift`.
+- Transcript message controls live in `QuillCodeTranscriptMessageView.swift`. User and assistant message bubbles, retry/use-as-draft actions, feedback actions, and the shared transcript copy button should evolve there; `WorkspaceSwiftUIView.swift` should only decide where timeline items are placed. Tool-card and artifact-preview rendering remains the next transcript family to extract.
 - Multi-agent collaboration should happen through pull requests and a serialized merge train rather than direct pushes to `main`. CI runs on `pull_request`, `push`, and `merge_group` events. The train watches `merge-train` and `automerge` labels, ignores draft or `do-not-merge` PRs, processes only the oldest eligible PR, updates behind branches before merging, and merges only after all non-train checks are successful. This keeps Codex/agent work reviewable while preserving a green mainline.
 - The first memories slice is auditable and user-authored. QuillCode loads bounded `.md`, `.txt`, and `.json` notes from `~/.quillcode/memories` and project `.quillcode/memories`, rejects symlinks and paths outside the memory root, truncates oversized notes with visible labels, snapshots loaded memories onto threads, injects them into TrustedRouter as non-command background context, and surfaces the count/sources through the top bar, Memories pane, static HTML renderer, Playwright harness, slash command, and macOS menu bar. Users can explicitly save global memories with `/remember text`, the Add memory command-palette action, or a model-authored `host.memory.remember` tool call when they ask QuillCode to remember a stable preference/fact; saved notes create transcript/tool audit records, refresh memory context immediately, and reject obvious credentials, tokens, passwords, and private keys. Global memories can be forgotten from the Memories pane and produce a transcript notice. Project memories stay read-only in-app because they are project-owned files. Richer redaction review, idle Chronicle jobs, editing UI, conflict resolution, and fully autonomous inference of memories without explicit user intent are separate safety/UX milestones.
